refactor: BaseEntity 분리
