+++ src/main/java/gift/product/infrastructure/persistence/ProductCategoryEntity.java
@@ -1,16 +1,12 @@
 package gift.product.infrastructure.persistence;
 
+import gift.core.BaseEntity;
 import gift.core.domain.product.ProductCategory;
 import jakarta.persistence.*;
 
 @Entity
 @Table(name = "product_category")
-public class ProductCategoryEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
+public class ProductCategoryEntity extends BaseEntity {
     @Column(name = "name", nullable = false, unique = true)
     private String name;
 
@@ -18,7 +14,7 @@ public class ProductCategoryEntity {
     }
 
     public ProductCategoryEntity(Long id, String name) {
-        this.id = id;
+        super(id);
         this.name = name;
     }
 
@@ -26,21 +22,15 @@ public class ProductCategoryEntity {
         this.name = name;
     }
 
-    public Long getId() {
-        return id;
-    }
-
     public String getName() {
         return name;
     }
 
     public ProductCategory toDomain() {
-        return new ProductCategory(id, name);
+        return new ProductCategory(getId(), name);
     }
 
     public static ProductCategoryEntity toEntity(ProductCategory productCategory) {
         return new ProductCategoryEntity(productCategory.id(), productCategory.name());
     }
-
-
 }
