+++ src/main/java/gift/product/infrastructure/persistence/ProductRepositoryImpl.java
@@ -2,6 +2,7 @@ package gift.product.infrastructure.persistence;
 
 import gift.core.PagedDto;
 import gift.core.domain.product.Product;
+import gift.core.domain.product.ProductCategory;
 import gift.core.domain.product.ProductRepository;
 import jakarta.annotation.Nonnull;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -77,7 +78,7 @@ public class ProductRepositoryImpl implements ProductRepository {
                 product.name(),
                 product.price(),
                 product.imageUrl(),
-                ProductCategoryEntity.toEntity(product.category())
+                ProductCategoryEntity.toEntity(ProductCategory.of(product.category()))
         );
     }
 }
